isolated update_func
